+++ src/components/ofxDatGuiGroups.h
@@ -53,6 +53,12 @@ class ofxDatGuiGroup : public ofxDatGuiButton {
             layout();
         }
     
+        void toggle()
+        {
+            mIsExpanded = !mIsExpanded;
+            layout();
+        }
+    
         void collapse()
         {
             mIsExpanded = false;
@@ -74,8 +80,6 @@ class ofxDatGuiGroup : public ofxDatGuiButton {
             if (mVisible){
                 ofPushStyle();
                 ofxDatGuiButton::draw();
-                ofSetColor(mIcon.color);
-                mImage->draw(x+mIcon.x, y+mIcon.y, mIcon.size, mIcon.size);
                 if (mIsExpanded) {
                     int mHeight = mStyle.height;
                     ofSetColor(mStyle.guiBackground, mStyle.opacity);
@@ -88,7 +92,12 @@ class ofxDatGuiGroup : public ofxDatGuiButton {
                         ofSetColor(mStyle.guiBackground, mStyle.opacity);
                         ofDrawRectangle(x, y+mHeight, mStyle.width, mStyle.vMargin);
                     }
+                    ofSetColor(mIcon.color);
+                    mIconOpen->draw(x+mIcon.x, y+mIcon.y, mIcon.size, mIcon.size);
                     for(int i=0; i<children.size(); i++) children[i]->drawColorPicker();
+                }   else{
+                    ofSetColor(mIcon.color);
+                    mIconClosed->draw(x+mIcon.x, y+mIcon.y, mIcon.size, mIcon.size);
                 }
                 ofPopStyle();
             }
@@ -133,7 +142,8 @@ class ofxDatGuiGroup : public ofxDatGuiButton {
         }
     
         int mHeight;
-        shared_ptr<ofImage> mImage;
+        shared_ptr<ofImage> mIconOpen;
+        shared_ptr<ofImage> mIconClosed;
         bool mIsExpanded;
     
 };
@@ -153,7 +163,8 @@ class ofxDatGuiFolder : public ofxDatGuiGroup {
         void setTheme(ofxDatGuiTheme* theme)
         {
             setComponentStyle(theme);
-            mImage = theme->icon.dropdown;
+            mIconOpen = theme->icon.groupOpen;
+            mIconClosed = theme->icon.groupClosed;
             setWidth(theme->layout.width, theme->layout.labelWidth);
         // reassign folder color to all components //
             for(auto i:children) i->setStripeColor(mStyle.stripe.color);
@@ -176,6 +187,18 @@ class ofxDatGuiFolder : public ofxDatGuiGroup {
         {
             if (buttonEventCallback != nullptr) {
                 buttonEventCallback(e);
+            }   else{
+                ofxDatGuiLog::write(ofxDatGuiMsg::EVENT_HANDLER_NULL);
+            }
+        }
+    
+        void dispatchToggleEvent(ofxDatGuiToggleEvent e)
+        {
+            if (toggleEventCallback != nullptr) {
+                toggleEventCallback(e);
+        // allow toggle events to decay into button events //
+            }   else if (buttonEventCallback != nullptr) {
+                buttonEventCallback(ofxDatGuiButtonEvent(e.target));
             }   else{
                 ofxDatGuiLog::write(ofxDatGuiMsg::EVENT_HANDLER_NULL);
             }
@@ -251,7 +274,7 @@ class ofxDatGuiFolder : public ofxDatGuiGroup {
         {
             ofxDatGuiToggle* toggle = new ofxDatGuiToggle(label, enabled);
             toggle->setStripeColor(mStyle.stripe.color);
-            toggle->onButtonEvent(this, &ofxDatGuiFolder::dispatchButtonEvent);
+            toggle->onToggleEvent(this, &ofxDatGuiFolder::dispatchToggleEvent);
             attachItem(toggle);
             return toggle;
         }
@@ -262,7 +285,7 @@ class ofxDatGuiFolder : public ofxDatGuiGroup {
             return slider;
         }
 
-        ofxDatGuiSlider* addSlider(string label, float min, float max, float val)
+        ofxDatGuiSlider* addSlider(string label, float min, float max, double val)
         {
             ofxDatGuiSlider* slider = new ofxDatGuiSlider(label, min, max, val);
             slider->setStripeColor(mStyle.stripe.color);
@@ -426,7 +449,8 @@ class ofxDatGuiDropdown : public ofxDatGuiGroup {
         void setTheme(ofxDatGuiTheme* theme)
         {
             setComponentStyle(theme);
-            mImage = theme->icon.dropdown;
+            mIconOpen = theme->icon.groupOpen;
+            mIconClosed = theme->icon.groupClosed;
             mStyle.stripe.color = theme->stripe.dropdown;
             setWidth(theme->layout.width, theme->layout.labelWidth);
         }