+++ history.md
@@ -1,6 +1,31 @@
 ### ofxDatGui Changelog
 
 --
+**v1.20** –– 04-07-16
+
+* fonts are now managed by [ofxSmartFont](https://github.com/braitsch/ofxSmartFont)
+* fix for [2dPad setValue & setPosition not working as expected](https://github.com/braitsch/ofxDatGui/issues/45)
+* fix for [folders not retaining color state after a theme is applied](https://github.com/braitsch/ofxDatGui/issues/36)
+* fix for [ambiguous placeholder conflict with boost](https://github.com/braitsch/ofxDatGui/issues/49)
+* fix for [sliders dispatching events on mousedown & mousedrag](https://github.com/braitsch/ofxDatGui/issues/63)
+* fix for [ofGetWidth & ofGetHeight returning incorrect values on retina displays](https://github.com/openframeworks/openFrameworks/pull/4858)
+* fix for [label components not respecting alignment settings](https://github.com/braitsch/ofxDatGui/commit/3bbb1591c01f87960e3a15aa62ea5ab16d10fa07)
+* fix for [scrollview not reseting itself properly after items deleted](https://github.com/braitsch/ofxDatGui/commit/9ab25457cac1d59975cc3e330beaf9f620970aef)
+* adding [gui->isMouseDown()](https://github.com/braitsch/ofxDatGui/issues/50)
+* adding [ofxDatGui::setAssetPath()](https://github.com/braitsch/ofxDatGui/issues/51) to allow for [configurable asset directory](https://github.com/braitsch/ofxDatGui/issues/53)
+* adding [folder->addSlider(ofParameter \<float\>)](https://github.com/braitsch/ofxDatGui/pull/43)
+* adding [gui->addFolder()](https://github.com/braitsch/ofxDatGui/pull/40)
+* adding [gui->setPosition(ofxDatGuiAnchor anchor)](https://github.com/braitsch/ofxDatGui/commit/eefc6bdfa79489fb9a152fa7b979a2192e560816)
+* adding [gui anchors BOTTOM LEFT & BOTTOM RIGHT](https://github.com/braitsch/ofxDatGui/pull/61)
+* adding [gui->collapse, gui->expand and gui->toggle](https://github.com/braitsch/ofxDatGui/issues/57)
+* adding [dropdown->toggle() & folder->toggle()](https://github.com/braitsch/ofxDatGui/commit/815c8fd4e678c0068430fb7a58555097deb958de)
+* [gui & component update loop optimizations](https://github.com/braitsch/ofxDatGui/commit/75ad03c6a0065a957c8105799b32717d83ba5344)
+* addToggle() now returns an [ofxDatGuiToggle instance instead of ofxDatGuiButton](https://github.com/braitsch/ofxDatGui/pull/43)
+* gui mouse events are now [detected before the app receives them](https://github.com/braitsch/ofxDatGui/issues/50)
+* slider->setPrecision() now allows [flag to not truncate precision of underlying value](https://github.com/braitsch/ofxDatGui/issues/42)
+* [all icons are now shared_ptrs](https://github.com/braitsch/ofxDatGui/pull/52) thanks [@mattfelsen!](https://github.com/mattfelsen)
+
+--
 **v1.10** –– 01-28-16
 
 * adding [scrollview](http://braitsch.github.io/ofxDatGui/index.html#scrollviews) component